# bismarck-game/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  bismarck-game_backend_internal_game_models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  bismarck-game_backend_internal_game_models.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  bismarck-game_backend_internal_game_models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  bismarck-game_backend_internal_game_models.UpdateUserRequest:
    properties:
      email:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
...
  internal_api_handlers.CheckEmergencyFuelRequest:
    properties:
      game_id:
        type: string
      unit_id:
        type: string
    type: object
  internal_api_handlers.CheckEmergencyFuelResponse:
    properties:
      current_fuel:
        type: integer
      emergency_turn:
        type: integer
      is_emergency_fuel:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Bismarck Game Team
  description: API для игры Bismarck Chase - пошаговой стратегической игры о морских
    сражениях
  title: Bismarck Game API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для смены пароля
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bismarck-game_backend_internal_game_models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Смена пароля
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bismarck-game_backend_internal_game_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Вход в систему
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Выход из системы
      tags:
      - Authentication
  /auth/profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получение профиля пользователя
      tags:
      - Authentication
    put:
      consumes:
      - application/json
      parameters:
      - description: Данные для обновления профиля
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bismarck-game_backend_internal_game_models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Обновление профиля пользователя
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bismarck-game_backend_internal_game_models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация пользователя
      tags:
      - Authentication
  /auth/validate:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Валидация токена
      tags:
      - Authentication
  /emergency-fuel/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для проверки аварийного топлива
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CheckEmergencyFuelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.CheckEmergencyFuelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Проверка и активация аварийного топлива
      tags:
      - Emergency Fuel
  /emergency-fuel/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID игры
        in: query
        name: game_id
        required: true
        type: string
      - description: ID юнита
        in: query
        name: unit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.CheckEmergencyFuelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение статуса аварийного топлива
      tags:
      - Emergency Fuel
...
securityDefinitions:
  Bearer:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"